refactor: extract formatWithBiome into shared tools module
Both generate-schema-types.ts and generate-registry-types.ts had the
same helper for spawning `biome check --write`. Move it into a single
tools/format-with-biome.ts module so both generators share one copy.
